Separate per-segment and whole-index document limits
Split IndexWriter s single ~2B cap into MAX_DOCS (int, per-segment,
unchanged) and MAX_TOTAL_DOCS (long, whole-index). The whole-index limit
is a deliberately bounded 2^34 (~16 billion), not Long.MAX_VALUE: a
segment stays capped at MAX_DOCS, so an unbounded total would be a trap
(unmergeable segments, meaningless accounting). IndexWriter keeps its doc
accounting; it just counts against the two limits.

+++ lucene/core/src/java/org/apache/lucene/index/DocumentsWriterPerThread.java
@@ -228,7 +228,7 @@ final class DocumentsWriterPerThread implements Accountable, Lock {
     if (n == 0) {
       return;
     }
-    final int maxDocs = IndexWriter.getActualMaxDocs();
+    final long maxDocs = IndexWriter.getActualMaxDocs();
     while (true) {
       long current = pendingNumDocs.get();
       long next = current + n;
